PR tweaks
