+++ ImperatorToCK3/Imperator/Characters/AccessoryGeneData.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using commonItems;
+
+namespace ImperatorToCK3.Imperator.Characters {
+	public class AccessoryGeneData {
+		public string geneName = "";
+		public string geneTemplate = "";
+		public string objectName = "";
+	}
+}
