+++ ImperatorToCK3/Imperator/Characters/PaletteCoordinates.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using commonItems;
+
+namespace ImperatorToCK3.Imperator.Characters {
+	public class PaletteCoordinates {
+		public uint x = 256; // palettes are 512x512
+		public uint y = 256;
+	}
+}
